Fix needs precision and improve reproduction system
- Change Needs storage from u8 to f32 to fix truncation bug preventing needs from increasing.
- Increase REPRODUCTION_RATE to 0.1 for more consistent population growth.
- Update reproduction_system to reset biological drives immediately when a pair meets, regardless of birth success.
- Implement birth_cooldown_system to properly manage post-birth recovery.
- Fix various type mismatches and UI bar rendering for high-precision needs.

+++ src/economy/transactions.rs
@@ -59,7 +59,7 @@ pub fn market_buy_system(
         }
 
         // Buy Food when hungry and have no food
-        if needs.hunger > HUNGER_BUY_THRESHOLD && agent.quantity_of(Good::Food) == 0 {
+        if needs.hunger > HUNGER_BUY_THRESHOLD as f32 && agent.quantity_of(Good::Food) == 0 {
             let (bought_item, qty) = market.try_buy_item(Good::Food, 1, &mut agent.currency);
             if let Some(it) = bought_item {
                 agent.add_item(it, qty as u8);
